Reworked screenshot mechanism.  Now screenshots are retrieved from video plugins with ReadScreen instead of CaptureScreen, and compressed/saved from main core instead of plugins.  Fixed 64-bit bug in Glide64.

+++ glide64/Gfx1.3.h
@@ -546,7 +546,7 @@ EXPORT void CALL DllConfig ( HWND hParent );
 EXPORT void CALL DllTest ( HWND hParent );
 
 
-EXPORT void CALL ReadScreen(void **dest, long *width, long *height);
+EXPORT void CALL ReadScreen(void **dest, int *width, int *height);
 
 /******************************************************************
   Function: DrawScreen
